model/requirements.txt
accelerate==1.2.1
huggingface-hub==0.27.0
mpmath==1.3.0
numpy==2.2.0
opencv-python==4.10.0.84
packaging==24.2
pillow==11.0.0
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scipy==1.14.1
sympy==1.13.1
tokenizers==0.21.0
torch==2.5.1
torchvision==0.20.1
transformers==4.47.1
